Breaking my own Location and Lexer
Future Jasmine please look into Location and Lexer.

Change location so that we don't require line or col.

If we need to print the msg out via reporter we split the thing by line,
have some data structure to output from the query source_start and source_end ->start and end of  line and col